Add files via upload
There is a film rental store that provides DVDs on a rental basis. The owner of this store is your friend and he is not at all computer savvy. To manage his business he has software that helps him to manage the business. 
He also has one helper who understands computers well and as a part-time job, he manages the software and helps the owner by providing different types of reports using SQL queries. 
Today, unfortunately, this helper is not available. So your friend has called you to help him with some of the reports that he needs to make important business decisions. 

Schema Information
Before you can help your friend it is very important for you to understand the schema. This schema contains a number of tables that relate to the various aspects of this store, such as the films available for rent, the customers who rent them, and the staff who manage the store and the rental process.
The tables in the database are interrelated, which allows them to run complex queries for data analysis.

Following is a brief description of a few of the tables 
Actor table contains information about the actors who have appeared in films, as well as information about the specific films in which they have appeared. 

The film table provides information about the films themselves, including their category, language, and other details. 
The customer Info table contains information about the customers who rent DVDs from the store, including their addresses and other personal information. 

Finally, the Store table contains information about the store itself, including details about the staff who work there, the rentals that are available, and the payments that are made by customers.